Refactor Rooms component, add Change room number functionality
