+++ src/components/Review.js
@@ -17,11 +17,13 @@ const modalStyle = {
   top: "50%",
   left: "50%",
   transform: "translate(-50%, -50%)",
-  width: 800,
+  width: "70%",
+  maxWidth: 800,
   bgcolor: "background.paper",
   border: "1px solid #000",
   boxShadow: 24,
   p: 4,
+  m: "auto",
 };
 
 export default function Review({ review }) {
